Fail Release workflow when `poetry version` exits with failure
The Release workflow detects a bogus previous version if the parent commit does
not contain a pyproject.toml file with a `tool.poetry.version` field.

Poetry is run as the first job in a pipeline, so its non-zero exit code is not
reported back to the action. Poetry writes its error message to stdout, which
results in a reported version "could" (second word of the error message).

Fix this by running the command via `bash -o pipefail -c`. Note that it is not
possible to use `set -o pipefail` inside the version command. The GitHub
Action (salsify/action-detect-and-tag-new-version) always uses /bin/sh, which
happens to be dash in Ubuntu. The dash shell does not support pipefail.